+++ configs/dataset/conll2003_base.yaml
@@ -1,7 +1,7 @@
 # This config on its own does nothing more than conll2003.yaml. However, it can be easily combined with further
 # preprocessing configs, see conll2003_select_n.yaml for an example.
 
-_target_: src.utils.config.execute_pipeline
+_target_: src.utils.execute_pipeline
 input:
   _target_: pytorch_ie.DatasetDict.load_dataset
   path: pie/conll2003
